Got getBlocks working
The JSON Marshall -> Unmarshall is working. Also fixed a bug that was
caused because the semver library only understands version strings of
the form "v<MAJOR>.<MINOR>.<PATCH>" and so it was treating my version
strings in format "<MAJOR>.<MINOR>.<PATCH>" without the "v" prefix as
invalid and comparing them to be the same. Annoying!